[#37]refactor: 패키지 수정 soma.haeya.edupi_user -> soma.edupi.user
