+++ src/main/java/soma/edupi/user/EdupiUserApplication.java
@@ -1,0 +1,13 @@
+package soma.edupi.user;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EdupiUserApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(EdupiUserApplication.class, args);
+    }
+
+}
