+++ src/main/java/soma/edupi/user/dto/request/MemberLoginRequest.java
@@ -1,0 +1,20 @@
+package soma.edupi.user.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class MemberLoginRequest {
+
+    @NotNull
+    @NotBlank
+    private String email;
+
+    @NotNull
+    @NotBlank
+    private String password;
+
+}
