+++ src/main/java/soma/edupi/user/exception/InnerServerException.java
@@ -1,0 +1,9 @@
+package soma.edupi.user.exception;
+
+public class InnerServerException extends RuntimeException {
+
+    public InnerServerException(String message) {
+        super(message);
+    }
+
+}
