--- src/main/java/soma/haeya/edupi_user/EdupiUserApplication.java
@@ -1,13 +1,0 @@
-package soma.haeya.edupi_user;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EdupiUserApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EdupiUserApplication.class, args);
-	}
-
-}
