--- src/main/java/soma/haeya/edupi_user/controller/MemberController.java
@@ -1,66 +1,0 @@
-package soma.haeya.edupi_user.controller;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CookieValue;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import soma.haeya.edupi_user.dto.TokenInfo;
-import soma.haeya.edupi_user.dto.request.MemberLoginRequest;
-import soma.haeya.edupi_user.dto.request.SignupRequest;
-import soma.haeya.edupi_user.dto.response.SignupResponse;
-import soma.haeya.edupi_user.service.MemberService;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/v1/member")
-public class MemberController implements MemberSpecification {
-
-    private final MemberService memberService;
-
-    @PostMapping("/login")
-    public ResponseEntity<Void> login(@Valid @RequestBody MemberLoginRequest memberLoginRequest,
-        HttpServletResponse response) {
-        String token = memberService.login(memberLoginRequest);
-
-        Cookie cookie = new Cookie("token", token);
-        cookie.setPath("/");
-        cookie.setHttpOnly(true);
-
-        response.addCookie(cookie);
-
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/login/info")
-    public ResponseEntity<TokenInfo> loginInfo(@CookieValue("token") String token) {
-        TokenInfo tokenInfo = memberService.findMemberInfo(token);
-
-        return ResponseEntity.ok(tokenInfo);
-    }
-
-    @PostMapping("/signup")
-    public ResponseEntity<SignupResponse> createPost(@Valid @RequestBody SignupRequest signupRequest)
-        throws JsonProcessingException {
-        return memberService.signUp(signupRequest);
-    }
-
-    @GetMapping("/logout")
-    public ResponseEntity<Void> logout(HttpServletResponse response) {
-        Cookie cookie = new Cookie("token", null);
-        cookie.setPath("/");
-        cookie.setMaxAge(0);
-
-        response.addCookie(cookie);
-
-        return ResponseEntity.ok().build();
-    }
-
-}
