--- src/main/java/soma/haeya/edupi_user/domain/Role.java
@@ -1,5 +1,0 @@
-package soma.haeya.edupi_user.domain;
-
-public enum Role {
-    ROLE_USER
-}
