--- src/main/java/soma/haeya/edupi_user/dto/response/ErrorResponse.java
@@ -1,15 +1,0 @@
-package soma.haeya.edupi_user.dto.response;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class ErrorResponse {
-
-    String message;
-
-    public ErrorResponse(String message) {
-        this.message = message;
-    }
-}
