--- src/main/java/soma/haeya/edupi_user/dto/response/SignupResponse.java
@@ -1,15 +1,0 @@
-package soma.haeya.edupi_user.dto.response;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class SignupResponse {
-
-    String message;
-
-    public SignupResponse(String message) {
-        this.message = message;
-    }
-}
