--- src/main/java/soma/haeya/edupi_user/exception/DbValidException.java
@@ -1,8 +1,0 @@
-package soma.haeya.edupi_user.exception;
-
-public class DbValidException extends RuntimeException {
-
-    public DbValidException(String message) {
-        super(message);
-    }
-}
