--- src/main/java/soma/haeya/edupi_user/exception/InnerServerException.java
@@ -1,9 +1,0 @@
-package soma.haeya.edupi_user.exception;
-
-public class InnerServerException extends RuntimeException {
-
-    public InnerServerException(String message) {
-        super(message);
-    }
-
-}
