+++ src/test/java/soma/edupi/user/EdupiUserApplicationTests.java
@@ -1,0 +1,13 @@
+package soma.edupi.user;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EdupiUserApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
