--- src/test/java/soma/haeya/edupi_user/EdupiUserApplicationTests.java
@@ -1,13 +1,0 @@
-package soma.haeya.edupi_user;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EdupiUserApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
